requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.11
django-cors-headers==2.0.2
djangocms-installer==0.9.6
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
packaging==16.8
psycopg2==2.7.1
PyJWT==1.4.2
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
tzlocal==1.3
